+++ src/CommandTable.cs
@@ -13,9 +13,11 @@ namespace EditorConfig
     internal sealed partial class PackageGuids
     {
         public const string guidEditorConfigPackageString = "6736c72d-7a27-49f0-9153-413ff47963ef";
+        public const string guidEditorFactoryString = "13cabf99-9eae-4ecc-a6f4-89f3ad54dc83";
         public const string guidEditorConfigPackageCmdSetString = "13cabf99-9eae-4ecc-a6f4-89f3ad54dc83";
         public const string guidWorkspaceExplorerToolWindowPackageCmdSetString = "cfb400f1-5c60-4f3c-856e-180d28def0b7";
         public static Guid guidEditorConfigPackage = new Guid(guidEditorConfigPackageString);
+        public static Guid guidEditorFactory = new Guid(guidEditorFactoryString);
         public static Guid guidEditorConfigPackageCmdSet = new Guid(guidEditorConfigPackageCmdSetString);
         public static Guid guidWorkspaceExplorerToolWindowPackageCmdSet = new Guid(guidWorkspaceExplorerToolWindowPackageCmdSetString);
     }
@@ -30,6 +32,7 @@ namespace EditorConfig
         public const int CreateEditorConfigFileId = 0x0100;
         public const int CreateEditorConfigFileAnyCodeId = 0x0200;
         public const int OpenSettingsId = 0x0300;
+        public const int NavigateToParentId = 0x0400;
         public const int idmAddNew = 0x2200;
     }
 }