+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Road map
 
 - [x] Fixed issue with formatter
+- [x] Added command to navigate to parent (F12 shortcut)
 - [ ] Swith icons to the official .editorconfig icon
 - [ ] Add item template for .editorconfig files